Move CLI related packages into internal
We don't want people to import these packages, so move these out into
private packages.
