+++ cmd/virtual-kubelet/internal/commands/root/tracing_register_jaeger.go
