+++ internal/manager/doc.go
