+++ internal/manager/resource.go
